Explain required GITHUB_TOKEN permissions
By default the `GITHUB_TOKEN` is given the required permissions to make
the API requests made by this action. If you set explicit permissions on
your token, however, the default permissions do not apply.

If you run the action without the appropriate permissions, you will see
an error message saying: "Error: Resource not accessible by integration"

This patch updates the README to explain that, if you are using explicit
permissions, you must specify the `actions:read` permission for the
action to work correctly.